+++ components/dropdown/dropdown-menu.ts
@@ -1,25 +1,20 @@
 import {Directive, ElementRef, HostBinding, Renderer} from '@angular/core';
 import {SuiDropdownService} from './dropdown.service';
-import {SuiTransition} from "../transition/transition";
+import {SuiTransition} from '../transition/transition';
 
 @Directive({
     selector: '[suiDropdownMenu]'
 })
-export class SuiDropdownMenu {
+export class SuiDropdownMenu extends SuiTransition {
     private _service:SuiDropdownService;
-    private _transition:SuiTransition;
 
     public set service(service:SuiDropdownService) {
         this._service = service;
         this._service.menuElement = this.el;
-        this._service.transition = this._transition;
-        if (service.isOpen) {
-            this._transition.isVisible = true;
-        }
+        this.setTransitionController(service.transition);
     }
 
-    public constructor(public el:ElementRef, public renderer:Renderer) {
-        this._transition = new SuiTransition(el, renderer);
-        this._transition.isVisible = false;
+    public constructor(public renderer:Renderer, public el:ElementRef) {
+        super(renderer, el);
     }
 }
