+++ components/index.ts
@@ -1,14 +1,14 @@
-export {SuiAccordionModule} from "./accordion/accordion.module";
-export {SuiCheckboxModule} from "./checkbox/checkbox.module";
-export {SuiCollapseModule} from "./collapse/collapse.module";
-export {SuiDimmerModule} from "./dimmer/dimmer.module";
-export {SuiDropdownModule} from "./dropdown/dropdown.module";
-export {SuiMessageModule} from "./message/message.module";
-export {SuiProgressModule} from "./progress/progress.module";
-export {SuiRatingModule} from "./rating/rating.module";
-export {SuiSearchModule} from "./search/search.module";
-export {SuiTabsModule} from "./tabs/tab.module";
-export {SuiSelectModule} from "./select/select.module";
-export {SuiTransitionModule} from "./transition/transition.module";
+export * from "./accordion/accordion.module";
+export * from "./checkbox/checkbox.module";
+export * from "./collapse/collapse.module";
+export * from "./dimmer/dimmer.module";
+export * from "./dropdown/dropdown.module";
+export * from "./message/message.module";
+export * from "./progress/progress.module";
+export * from "./rating/rating.module";
+export * from "./search/search.module";
+export * from "./tabs/tab.module";
+export * from "./select/select.module";
+export * from "./transition/transition.module";
 
-export {SuiModule} from "./sui.module";
+export * from "./sui.module";
