+++ components/search/search.ts
@@ -94,7 +94,7 @@ export class SuiSearch implements AfterViewInit {
 
     @HostBinding('class.visible')
     public get isVisible() {
-        return this._dropdownService.isVisible;
+        return this._dropdownService.transition.isVisible;
     }
 
     @HostBinding('class.active')
