+++ components/transition/transition.ts
@@ -1,16 +1,187 @@
-import {Directive, HostBinding, ElementRef, AfterViewInit, Renderer} from "@angular/core";
+import {Renderer, ElementRef, Directive, Input, HostBinding} from '@angular/core';
 
-export interface ISuiAnimation {
-    name:string;
-    duration?:number;
-    display?:string;
-    direction?:string;
-    callback?:() => any;
+export enum TransitionDirection {
+    In,
+    Out,
+    Either,
+    Static
 }
 
-interface ISuiConfiguredAnimation extends ISuiAnimation {
-    classes?:string[];
-    static?:boolean;
+export class Transition {
+    public readonly type:string;
+
+    public readonly duration:number;
+
+    public direction:TransitionDirection;
+
+    public get directionClass() {
+        switch (this.direction) {
+            case TransitionDirection.In: return "in";
+            case TransitionDirection.Out: return "out";
+        }
+    }
+
+    public readonly classes:string[];
+
+    public onComplete:() => any;
+
+    constructor(name:string, duration:number = 250, direction:TransitionDirection = TransitionDirection.Either, onComplete:(() => any) = () => {}) {
+        this.type = name;
+        this.duration = duration;
+        this.direction = direction;
+        this.classes = this.type.split(" ");
+        this.onComplete = onComplete;
+    }
+}
+
+export class TransitionController {
+    private _renderer:Renderer;
+
+    private _element:ElementRef;
+
+    private get _isReady() {
+        return this._renderer != null && this._element != null;
+    }
+
+    private _display:string;
+
+    private _queue:Transition[];
+
+    private _isAnimating:boolean;
+
+    public get isAnimating() {
+        return this._isAnimating;
+    }
+
+    private _isVisible:boolean;
+
+    public get isVisible() {
+        return this._isVisible;
+    }
+
+    private _isHidden:boolean;
+
+    public get isHidden() {
+        return this._isHidden;
+    }
+    
+    private get _queueFirst() {
+        return this._queue[0];
+    }
+
+    private get _queueLast() {
+        return this._queue[this._queue.length - 1];
+    }
+
+    private _animationTimeout:any;
+
+    constructor(isInitiallyVisible:boolean = true, display:string = "block") {
+        this._isVisible = isInitiallyVisible;
+        this._isHidden = !this._isVisible;
+
+        this._display = display;
+        this._queue = [];
+
+        this._isAnimating = false;
+    }
+
+    public registerRenderer(renderer:Renderer) {
+        this._renderer = renderer;
+        this.performTransition();
+    }
+
+    public registerElement(element:ElementRef) {
+        this._element = element;
+        this.performTransition();
+    }
+
+    public animate(transition:Transition) {
+        let isDirectionless = ["jiggle", "flash", "shake", "pulse", "tada", "bounce"].indexOf(transition.type) != -1;
+        if (isDirectionless) {
+            transition.direction = TransitionDirection.Static;
+        }
+        else if (!transition.direction || transition.direction == TransitionDirection.Either) {
+            transition.direction = this._isVisible ? TransitionDirection.Out : TransitionDirection.In;
+            if (this._queueLast) {
+                transition.direction = this._queueLast.direction == TransitionDirection.In ? TransitionDirection.Out : TransitionDirection.In;
+            }
+        }
+
+        this._queue.push(transition);
+
+        this.performTransition();
+    }
+
+    private performTransition() {
+        if (!this._isReady || this._isAnimating || !this._queueFirst) {
+            return;
+        }
+
+        this._isAnimating = true;
+
+        let transition = this._queueFirst;
+
+        transition.classes.forEach(c => this._renderer.setElementClass(this._element, c, true));
+        this._renderer.setElementClass(this._element, `animating`, true);
+        this._renderer.setElementClass(this._element, transition.directionClass, true);
+
+        this._renderer.setElementStyle(this._element, `animationDuration`, `${transition.duration}ms`);
+        this._renderer.setElementStyle(this._element, `display`, this._display);
+
+        if (transition.direction == TransitionDirection.In) {
+            this._isHidden = false;
+        }
+    
+        this._animationTimeout = setTimeout(() => this.finishTransition(transition), transition.duration);
+    }
+
+    private finishTransition(transition:Transition) {
+        transition.classes.forEach(c => this._renderer.setElementClass(this._element, c, false));
+        this._renderer.setElementClass(this._element, `animating`, false);
+        this._renderer.setElementClass(this._element, transition.directionClass, false);
+
+        this._renderer.setElementStyle(this._element, `animationDuration`, null);
+        this._renderer.setElementStyle(this._element, `display`, null);
+
+        if (transition.direction == TransitionDirection.In) {
+            this._isVisible = true;
+        }
+        else if (transition.direction == TransitionDirection.Out) {
+            this._isVisible = false;
+            this._isHidden = true;
+        }
+
+        if (transition.onComplete) {
+            transition.onComplete();
+        }
+
+        this._queue.shift();
+        this._isAnimating = false;
+
+        this.performTransition();
+    }
+
+    public stop(transition:Transition = this._queueFirst) {
+        if (!transition || !this._isAnimating) {
+            return;
+        }
+
+        clearTimeout(this._animationTimeout);
+        this.finishTransition(transition);
+    }
+
+    public stopAll() {
+        this.clearQueue();
+        this.stop();
+    }
+
+    public clearQueue() {
+        if (this.isAnimating) {
+            this._queue = [this._queueFirst];
+            return;
+        }
+        this._queue = [];
+    }
 }
 
 @Directive({
@@ -18,157 +189,37 @@ interface ISuiConfiguredAnimation extends ISuiAnimation {
     exportAs: 'transition'
 })
 export class SuiTransition {
-    private _isAnimating = false;
+    private _transitionController:TransitionController;
 
-    public get isAnimating() {
-        return this._isAnimating;
+    @Input()
+    public set suiTransition(tC:TransitionController) {
+        this.setTransitionController(tC);
     }
 
-    public set isAnimating(value:boolean) {
-        this._isAnimating = value;
-        this.renderer.setElementClass(this.el.nativeElement, "animating", value);
-    }
+    @HostBinding('class.transition')
+    public transitionClass = true;
 
-    private animationTimeout:any;
-
-    private _isVisible:boolean = null;
-
+    @HostBinding('class.visible')
     public get isVisible() {
-        return this._isVisible;
+        if (this._transitionController) {
+            return this._transitionController.isVisible;
+        }
+        return false;
     }
 
-    public set isVisible(value:boolean) {
-        this._isVisible = value;
-        this.renderer.setElementClass(this.el.nativeElement, "visible", value);
-        this.isHidden = this.isVisible !== null && !this.isVisible && !this.isAnimating;
-    }
-
-    private _isHidden: boolean;
-
+    @HostBinding('class.hidden')
     public get isHidden() {
-        return this._isHidden;
+        if (this._transitionController) {
+            return this._transitionController.isHidden;
+        }
+        return false;
     }
 
-    public set isHidden(value:boolean) {
-        this._isHidden = value;
-        this.renderer.setElementClass(this.el.nativeElement, "hidden", value);
-    }
+    constructor(public renderer:Renderer, public element:ElementRef) {}
 
-    private queue:ISuiConfiguredAnimation[] = [];
-
-    constructor(private el:ElementRef, private renderer:Renderer) {
-        this.renderer.setElementClass(this.el.nativeElement, "transition", true);
-
-        // Set initial state to visible / invisible
-        setTimeout(() => {
-            let style = window.getComputedStyle(this.el.nativeElement);
-            if (this.isVisible === null) {
-                this.isVisible = style.display !== 'none';
-            }
-        });
-    }
-
-    private queueFirst() {
-        return this.queue.slice(0, 1).pop();
-    }
-
-    private queueLast() {
-        return this.queue.slice(-1).pop();
-    }
-
-    public animate(animation:ISuiAnimation) {
-        animation = Object.assign({}, animation);
-        let anim = animation as ISuiConfiguredAnimation;
-
-        anim.classes = animation.name.split(" ");
-        if (!anim.duration) {
-            anim.duration = 250;
-        }
-        if (!anim.display) {
-            anim.display = 'block';
-        }
-        if (!anim.static) {
-            anim.static = ["jiggle", "flash", "shake", "pulse", "tada", "bounce"].indexOf(anim.name) != -1;
-        }
-        if (!anim.direction) {
-            anim.direction = this.isVisible ? "out" : "in";
-            let queueLast = this.queueLast();
-            if (queueLast) {
-                anim.direction = queueLast.direction == "in" ? "out" : "in"
-            }
-        }
-
-        this.queue.push(animation);
-
-        this.performAnimation();
-    }
-
-    private performAnimation() {
-        if (this.isAnimating) {
-            return;
-        }
-        let animation = this.queue.slice(0, 1).pop();
-        if (!animation) {
-            return;
-        }
-
-        this.isAnimating = true;
-        this.isVisible = true;
-        this.isHidden = false;
-
-        animation.classes.forEach(c => this.renderer.setElementClass(this.el.nativeElement, c, true));
-        this.renderer.setElementClass(this.el.nativeElement, animation.direction, true);
-        this.renderer.setElementStyle(this.el.nativeElement, `animationDuration`, `${animation.duration}ms`);
-        if (animation.direction == "in") {
-            this.renderer.setElementStyle(this.el.nativeElement, `display`, animation.display);
-        }
-
-        this.animationTimeout = setTimeout(() => this.finishAnimation(animation), animation.duration);
-    }
-
-    private finishAnimation(animation:ISuiConfiguredAnimation) {
-        this.isAnimating = false;
-        animation.classes.forEach(c => this.renderer.setElementClass(this.el.nativeElement, c, false));
-        this.renderer.setElementClass(this.el.nativeElement, animation.direction, false);
-        this.renderer.setElementStyle(this.el.nativeElement, `animationDuration`, null);
-        this.renderer.setElementStyle(this.el.nativeElement, `display`, null);
-
-        this.isVisible = animation.direction == "in" ? true : false;
-        if (animation.static) {
-            this.isVisible = !this.isVisible;
-        }
-
-        if (animation.callback) {
-            animation.callback();
-        }
-
-        this.queue.shift();
-
-        this.performAnimation();
-    }
-
-    public stop() {
-        if (this.isAnimating) {
-            clearTimeout(this.animationTimeout);
-            this.finishAnimation(this.queueFirst());
-        }
-    }
-
-    public stopAll() {
-        if (this.isAnimating) {
-            clearTimeout(this.animationTimeout);
-            this.finishAnimation(this.queueFirst());
-        }
-        this.clearQueue();
-    }
-
-    public clearQueue() {
-        if (this.isAnimating) {
-            this.queue = [this.queueFirst()];
-            return;
-        }
-        this.queue = [];
+    public setTransitionController(transitionController:TransitionController) {
+        this._transitionController = transitionController;
+        this._transitionController.registerRenderer(this.renderer);
+        this._transitionController.registerElement(this.element.nativeElement);
     }
 }
-
-export const SUI_TRANSITION_DIRECTIVES = [SuiTransition];
