+++ demo/src/app/pages/test/test.page.html
@@ -7,7 +7,7 @@
 <demo-page-content>
     <h2 class="ui dividing header">Examples</h2>
     <div class="ui segment">
-        <img id="testImage" src="http://semantic-ui.com/images/leaves/1.png" class="ui autumn leaf image" [newSuiTransition]="transition">
+        <img id="testImage" src="http://semantic-ui.com/images/leaves/1.png" class="ui autumn leaf image" [suiTransition]="transition">
         <button class="ui button" (click)="animate()">Animate!</button>
         <button class="ui button" (click)="stop()">Stop!</button>
         <button class="ui button" (click)="stopAll()">Stop all!</button>
