+++ demo/src/app/pages/test/test.page.ts
@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import {SuiTransition} from "../../../../../components/transition/transition";
-import {TransitionController, Transition} from '../../../../../components/transition/new-transition';
+import {TransitionController, Transition} from '../../../../../components/transition/transition';
 
 @Component({
   selector: 'demo-page-test',
